pyproject.toml
[project]
name = "extracredit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "edapi>=0.1.0",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
